src/test/java/helix/MatrixHelixMorphTest.java
package helix;

import org.junit.Test;

import static org.junit.Assert.assertTrue;

public class MatrixHelixMorphTest {

    /** Check if two integer matrices have the same contents
     *
     * @param matrix1
     * @param matrix2
     * @return true if the contents of matrix1 and matrix2 are identical and false otherwise
     */
    public static boolean testEqual(int[][] matrix1, int[][] matrix2) {

        /*
            Basic sanity checking:
                Verify that lengths are equal
         */
        if (matrix1 == null || matrix2 == null) {
            return true;
        }
        if (matrix1.length != matrix2.length) {
            return false;
        }
        if (matrix1[0].length != matrix2[0].length) {
            return false;
        }

        /*
            Test the contents of the matrices
         */
        for (int i = 0; i < matrix1.length; i++) {
            for (int j = 0; j < matrix1[0].length; j++) {
                if (matrix1[i][j] != matrix2[i][j]) {
                    return false;
                }
            }
        }

        return true;
    }

    @Test
    public void test2x2() {
        int[][] inMatrix  = { {1, 2}, {3, 4} };
        int[][] expOutput = { {1, 2}, {4, 3} };
        assertTrue(testEqual(expOutput, MatrixHelixMorph.helix(inMatrix)));
    }

    @Test
    public void test1x1() {
        int[][] inMatrix  = { {1} };
        int[][] expOutput = { {1} };
        assertTrue(testEqual(expOutput, MatrixHelixMorph.helix(inMatrix)));
    }

    // ADD TESTS TO CHECK PARTITIONS

    @Test
    public void test3x3() {
        int[][] inMatrix = {{1, 2, 3},
                            {4, 5, 6},
                            {7, 8, 9},};

        int[][] expOutput = {{1, 2, 3},
                             {8, 9, 4},
                             {7, 6, 5},};
        assertTrue(testEqual(expOutput, MatrixHelixMorph.helix(inMatrix)));
    }

    @Test
    public void test4x4() {
        int[][] inMatrix = {{3, 6, 9, 12},
                            {15, 18, 21, 24},
                            {27, 30, 33, 36},
                            {39, 42, 45, 48}};

        int[][] expOutput = {{3, 6, 9, 12},
                             {36, 39, 42, 15},
                             {33, 48, 45, 18},
                             {30, 27, 24, 21}};
        assertTrue(testEqual(expOutput, MatrixHelixMorph.helix(inMatrix)));
    }

    @Test
    public void test5x4() {
        int[][] inMatrix = {{7, 5, 1, 4, 6},
                            {3, 2, 9, 1, 8},
                            {8, 3, 2, 4, 1},
                            {10, 11, 15, 19, 14}};

        int[][] expOutput = {{7, 5, 1, 4, 6},
                            {4, 1, 10, 11, 3},
                            {2, 14, 19, 15, 2},
                            {3, 8, 8, 1, 9}};
        assertTrue(testEqual(expOutput, MatrixHelixMorph.helix(inMatrix)));
    }
}
